Switched from GET to POST
I liked using GET because you could bookmark URLs for specific
listings. However, it's not a good choice because if you run,
say, an `add` or a `do` in Mobile Safari, switch to another app or
browser tab and then come back later, the page may reload on you
and cause a repeat of your previous action, with potentially
bad consequences. So I switched to POST.