Fixed regex so lsc and lsprj work properly.
In the last commit (a2f785d) I added '\s' to the beginning of the
preg_replace pattern so that 'foo@bar.com' wouldn't match, but this
meant (unnumbered) output of lsc and lsprj wouldn't match either. Now
the pattern matches any @ or + word with leading space or at the start
of a line.